Add files via upload
This project provides an opportunity to play against the computer. The computer will think of a random number from 1 to any bigger number, and ask you to guess it. The computer will tell you if each guess is too high or too low. This is a good game to code because it uses random numbers, loops, and input from the user in a short program. To me, this is a mind-practicing activity because me, as the player I try to guess the computer's secret number. The computer will then need to store the secret number in a variable so that it can compare it to my guesses. Another objective I thought is that this game can also give you a chance of what it is like in the hacking/cybersecurity space. Because there is a lot of guessing going on there, with the tools of executing tasks.